requirements.txt
Flask==3.0.0
spotipy==2.23.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.1.2
gunicorn==21.2.0
APScheduler==3.10.4
Flask-Limiter==3.5.0
